refactor: Alterando a estrutura de páginas.
